src/test/java/com/example/inflearnbasic/discount/RateDiscountPolicyTest.java
package com.example.inflearnbasic.discount;

import com.example.inflearnbasic.annotation.MainDiscountPolicy;
import com.example.inflearnbasic.memeber.Grade;
import com.example.inflearnbasic.memeber.Member;
import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

class RateDiscountPolicyTest {

    RateDiscountPolicy discountPolicy = new RateDiscountPolicy();

    @Test
    @DisplayName("VIP는 10% 할인이 적용되어야 한다") // 테스트코드 작동시 이름 뜸
    void vip_o() {
        //given
        Member member = new Member(1L,"memberVIP", Grade.VIP);
        //when
        int discount = discountPolicy.discount(member, 10000);
        //then
        Assertions.assertThat(discount).isEqualTo(1000);
    }

    @Test  //이건 실패테스트코드 임 위에처럼 성공만 보지말고 실패코드도 해놔야 안심
    @DisplayName("VIP가 아니면 할인이 적용되지 않아야 한다")
    void vip_x() {
        //given
        Member member = new Member(2L, "memeberBASIC", Grade.BASIC);
        //when
        int discount = discountPolicy.discount(member, 10000);
        //then
        Assertions.assertThat(discount).isEqualTo(1000);
    }


}